multifile: Fix saver expecting an existing file
The format saver previously expected a file to already exist at a given save path, so that it could restore its contents should it fail to save.

Now, the format saver will delete a file if it was just created and could not be saved (rather than throwing a File Not Found error).