bam: Implement UserVertexTransform
Previously the only type of vertex transform for transform blends we were handling was JointVertexTransforms, which nicely aligned with a Panda3D Character -> Godot Skeleton3D setup. However, other types of vertex transforms exist can be found in BAM files, particularly UserVertexTransforms, which supplies a static arbitrary matrix.

Whenever possible, we try to use Godot's in-engine functionality to replicate Panda3D features, to leverage the C++ backend. So as a bit of reuse, we'll look for any UserVertexTransforms that are referenced by a Character, and add a "static" bone to the resulting Skeleton3D that has the supplied transform as its bone rest.

This isn't a complete solution, however. While JointVertexTransforms were only sensible under a Character object, ALL Panda3D geometry can have an associate TransformBlendTable, and so technically any mesh should be able to handle a UserVertexTransform. If it's a big enough issue we can tackle it later on, but the Panda3D documentation suggests UserVertexTransforms shouldn't appear too commonly as they're mainly used for testing.