Add config option to handle excess transform blends
Panda3D does not have a limit on the number of transform blends (bone weights) that can influence a vertex. Godot has a hard limit of up to 8 bone weights per vertex.

This commit adds an import/config option to change the behavior when encountering a model with more than 8 transform blends. The default is to call a parse_error() as before, but you can now choose to drop the least significant blends/weights, with or without a warning.